# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stickers
    description: Sticker generation and management
  - name: openai
    description: OpenAI AI image operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /stickers:
    get:
      operationId: listStickers
      tags: [stickers]
      summary: List all generated stickers
      responses:
        "200":
          description: List of stickers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sticker"
    post:
      operationId: createSticker
      tags: [stickers]
      summary: Save a generated sticker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStickerBody"
      responses:
        "201":
          description: Created sticker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sticker"
  /stickers/{id}:
    delete:
      operationId: deleteSticker
      tags: [stickers]
      summary: Delete a sticker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /stickers/recent:
    get:
      operationId: getRecentStickers
      tags: [stickers]
      summary: Get most recently generated stickers
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent stickers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sticker"
  /openai/generate-image:
    post:
      operationId: generateOpenaiImage
      tags: [openai]
      summary: Generate a sticker image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateOpenaiImageBody"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateOpenaiImageResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Sticker:
      type: object
      properties:
        id:
          type: integer
        prompt:
          type: string
        imageData:
          type: string
          description: Base64 encoded image data
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - prompt
        - imageData
        - createdAt
    CreateStickerBody:
      type: object
      properties:
        prompt:
          type: string
        imageData:
          type: string
      required:
        - prompt
        - imageData
    GenerateOpenaiImageBody:
      type: object
      properties:
        prompt:
          type: string
        size:
          type: string
          enum: ["1024x1024", "512x512", "256x256"]
      required:
        - prompt
    GenerateOpenaiImageResponse:
      type: object
      properties:
        b64_json:
          type: string
      required:
        - b64_json
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error