# lib/api-zod/src/generated/types/sticker.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Sticker {
  id: number;
  prompt: string;
  /** Base64 encoded image data */
  imageData: string;
  createdAt: Date;
}